Convert GuestProperties to a function component
