+++ web/html/src/manager/virtualization/guests/create/guests-create.js
@@ -7,7 +7,7 @@ const { hot } = require('react-hot-loader');
 const { TopPanel } = require('components/panels/TopPanel');
 const MessagesUtils = require('components/messages').Utils;
 const { getOrderedItemsFromModel } = require('components/input/FormMultiInput');
-const { GuestProperties } = require('../guest-properties');
+const { GuestProperties } = require('../GuestProperties');
 const { VirtualizationGuestActionApi } = require('../virtualization-guest-action-api');
 const GuestNicsPanel = require('../properties/guest-nics-panel');
 const DiskUtils = require('../properties/disk-utils');
