+++ web/html/src/manager/virtualization/guests/edit/guests-edit.js
@@ -10,7 +10,7 @@ const { TopPanel } = require('components/panels/TopPanel');
 const MessagesUtils = require('components/messages').Utils;
 const { Loading } = require('components/utils/Loading');
 const { getOrderedItemsFromModel } = require('components/input/FormMultiInput');
-const { GuestProperties } = require('../guest-properties');
+const { GuestProperties } = require('../GuestProperties');
 const GuestNicsPanel = require('../properties/guest-nics-panel');
 const DiskUtils = require('../properties/disk-utils');
 const { VirtualizationGuestActionApi } = require('../virtualization-guest-action-api');
@@ -102,8 +102,7 @@ class GuestsEdit extends React.Component<Props> {
                   const onSubmit = properties => onAction('update', [this.props.guestUuid],
                     GuestsEdit.getRequestParameterFromModel(properties, initialModel));
                   const messages = [].concat(definitionMessages, actionMessages)
-                    .filter(item => item)
-                    .map(item => MessagesUtils.error(item))[0];
+                    .filter(item => item);
                   const guestName = definition !== null ? definition.name : '';
                   return (
                     <TopPanel title={guestName} icon="fa spacewalk-icon-virtual-guest">
